+++ torchbenchmark/models/nanogpt_generate/metadata.yaml
@@ -1,5 +1,7 @@
 eval_benchmark: false
 eval_deterministic: false
 eval_nograd: true
+not_implemented:
+- jit: true
 train_benchmark: false
 train_deterministic: false
